test(e2e): add fake-upstream OTLP smoke
Move the in-process OTLP HTTP receiver into the shared e2e harness so it is available without the Ollama tag.

Add a fake-upstream e2e that starts the real Hecate binary with OTLP traces and metrics enabled, then asserts gateway.provider spans and hecate.provider.calls metrics reach the sink.

Keep the Ollama suite on the same shared sink while preserving its real-runtime trace and metric smoke coverage.
